backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-ollama>=0.2.0
langgraph>=0.2.0

neo4j==5.26.0

pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

fpdf2==2.8.2
networkx==3.4.2

httpx>=0.27.0
